new Event system / general cleanup

+++ src/engine/console.cpp
@@ -14,6 +14,7 @@
 #include "engine/ui.h"
 #include "console.h"
 #include <list>
+#include <algorithm>
 
 //#define MAXCONLINES 1000
 
@@ -879,6 +880,7 @@ namespace {
         return color;
     }
 }
+
 void Console::Update()
 {
     if (m_BufferedWidth == 0.0f && m_BufferedHeight == 0.0f)
@@ -889,7 +891,7 @@ void Console::Update()
             .GetFramebufferSize(m_BufferedWidth, m_BufferedHeight);
     }
 
-    auto lineNum = std::max(0ull, m_ConsoleLines.size() - m_Config.Regular.LineNum);
+    auto lineNum = std::max(0ul, (unsigned long)m_ConsoleLines.size() - m_Config.Regular.LineNum);
     auto height = (engine::nui::GetDevice().GetLineHeight() + 1.0f) * std::min((int)m_ConsoleLines.size(), m_Config.Regular.LineNum) + 2.0f;
     nk_style_push_style_item(
         engine::nui::GetNKContext(),
